requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2020.6.20
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
wincertstore==0.2
